# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OTP Management OpenApi",
        "description": "OTP Management API Explorer.",
        "contact": {
            "email": "gladwell_n@live.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api/v1",
            "description": "Swagger OpenApi server"
        }
    ],
    "paths": {
        "/auth/otp/request": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Request OTP.",
                "description": "Sends an OTP to the user's email address.",
                "operationId": "requestOtp",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "email"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "gladwell_n@live.com"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OTP sent"
                    },
                    "403": {
                        "description": "Maximum OTP requests reached."
                    },
                    "422": {
                        "description": "Invalid email address"
                    },
                    "429": {
                        "description": "Too Many Requests"
                    }
                }
            }
        },
        "/auth/otp/validate": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Validate OTP.",
                "description": "Validates the OTP provided by the user.",
                "operationId": "validateOtp",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "email"
                                    },
                                    "pin": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "gladwell_n@live.com",
                                    "identifier": 1,
                                    "pin": 123456
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OTP valid"
                    },
                    "403": {
                        "description": "Invalid OTP"
                    },
                    "400": {
                        "description": "No active OTP"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "429": {
                        "description": "Too Many Requests"
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}